PEP 517 package structure and config.
Also removed the need of a config file. Everything will be
inferred from args and the like. The goal of the project right
now is to be able to stand on top of a user's repository and
running:
$ mlflow-webhook deploys
to listen for webhooks and run pipelines listed in the branch
`deploys`. No config files, just args.
check-changes was added as a script so it could run directly in the env
without needing to find the correct python version. The user just needs
to activate their conda env (this requirement is due to MLflow, not to
us), install mlflow-webhook and run it.